# openapi.yaml
openapi: 3.0.0
info:
  title: apiTienda
  version: '1.1'
  description: API para gestionar una tienda en línea, incluyendo funcionalidades para manejar usuarios, productos, categorías, pedidos y detalles de pedidos. Permite realizar operaciones CRUD en estas entidades, facilitando la gestión completa de un sistema de comercio electrónico.
servers:
  - url: http://localhost:8080
    description: Servidor local
paths:
  /categories:
    get:
      summary: Obtener todas las categorías
      parameters:
        - in: query
          name: name
          schema:
            type: string
        - in: query
          name: active
          schema:
            type: boolean
        - in: query
          name: creationDate
          schema:
            type: string
      responses:
        '200':
          description: Lista de categorías obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryOutDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      summary: Añadir una nueva categoría
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '201':
          description: Categoría creada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /categories/{categoryId}:
    get:
      summary: Obtener una categoría por ID
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
            format: long
      responses:
        '200':
          description: Categoría obtenida con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '404':
          $ref: '#/components/responses/CategoryNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      summary: Actualizar una categoría
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Categoría actualizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/CategoryNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Actualizar una categoría
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Category'
      responses:
        '200':
          description: Categoría actualizada con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/CategoryNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Eliminar una categoría
      parameters:
        - in: path
          name: categoryId
          required: true
          schema:
            type: integer
            format: long
      responses:
        '204':
          description: Categoría eliminada con éxito
        '404':
          $ref: '#/components/responses/CategoryNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /orders:
    get:
      summary: Obtener todos los pedidos
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: paymentMethod
          schema:
            type: string
        - in: query
          name: creationDate
          schema:
            type: string
      responses:
        '200':
          description: Lista de pedidos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderOutDTO'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /orders/{orderId}:
    get:
      summary: Obtener un pedido por ID
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
      responses:
        '200':
          description: Pedido obtenido con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      summary: Eliminar un pedido
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
      responses:
        '204':
          description: Pedido eliminado con éxito
        '404':
          $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/orders:
    post:
      summary: Añadir un nuevo pedido para un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '201':
          description: Pedido creado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      summary: Obtener pedidos de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
      responses:
        '200':
          description: Lista de pedidos obtenida con éxito
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OrderOutDTO'
        '404':
          $ref: '#/components/responses/UserNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/orders/{orderId}:
    put:
      summary: Actualizar un pedido de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Pedido actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Usuario o pedido no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    patch:
      summary: Actualizar un pedido de un usuario
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: integer
            format: long
        - in: path
          name: orderId
          required: true
          schema:
            type: integer
            format: long
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
      responses:
        '200':
          description: Pedido actualizado con éxito
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          description: Usuario o pedido no encontrado
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/responses/UserNotFound'
                  - $ref: '#/components/responses/OrderNotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    Category:
      type: object
      properties:
        id:
          type: integer
          format: long
        name:
          type: string
        description:
          type: string
        creationDate:
          type: string
          format: date
        active:
          type: boolean
        image:
          type: string
      example:
        id: 1
        name: "Electrónica"
        description: "Productos electrónicos y gadgets"
        creationDate: "2023-01-15"
        active: true
        image: "https://example.com/images/electronics.jpg"
    CategoryOutDTO:
      type: object
      properties:
        id:
          type: integer
          format: long
        name:
          type: string
        description:
          type: string
        active:
          type: boolean
      example:
        id: 2
        name: "Ropa"
        description: "Moda y accesorios"
        active: true
...
    Order:
      type: object
      properties:
        id:
          type: integer
          format: long
        status:
          type: string
        total:
          type: number
          format: float
        address:
          type: string
        creationDate:
          type: string
          format: date
        paymentMethod:
          type: string
      example:
        id: 5001
        status: "En proceso"
        total: 1299.98
        address: "Calle Comercio 78, Ciudad"
        creationDate: "2025-01-20"
        paymentMethod: "Tarjeta de crédito"
    OrderOutDTO:
      type: object
      properties:
        id:
          type: integer
          format: long
        status:
          type: string
        total:
          type: number
          format: float
        userId:
          type: integer
          format: long
        address:
          type: string
        creationDate:
          type: string
        paymentMethod:
          type: string
      example:
        id: 5002
        status: "Enviado"
        total: 799.99
        userId: 1001
        address: "Avenida Principal 90, Pueblo"
        creationDate: "2025-01-21"
        paymentMethod: "PayPal"
...
    CategoryNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "Categoría no encontrada"
      example:
        errorcode: 404
        message: "Categoría no encontrada"
...
    OrderNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "Pedido no encontrado"
      example:
        errorcode: 404
        message: "Pedido no encontrado"
...
    UserNotFoundException:
      type: object
      properties:
        errorcode:
          type: integer
          example: 404
        message:
          type: string
          example: "El usuario no existe"
      example:
        errorcode: 404
        message: "El usuario no existe"
    ErrorResponse:
      type: object
      properties:
        errorcode:
          type: integer
        message:
          type: string
        errorMessages:
          type: object
          additionalProperties:
            type: string
      example:
        errorcode: 500
        message: "Internal Server Error"
  responses:
   BadRequest:
    description: Solicitud inválida
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/ErrorResponse'
        examples:
          BadRequest:
            value:
              errorcode: 400
              message: "Bad Request"
          ValidationError:
            value:
              errorcode: 400
              message: "Validation Error"
              errorMessages:
                name: "El nombre es requerido"
                email: "El formato del email es inválido"
   CategoryNotFound:
    description: Categoría no encontrada
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/CategoryNotFoundException'
...
   UserNotFound:
    description: El usuario no existe
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/UserNotFoundException'
   OrderNotFound:
    description: Pedido no encontrado
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/OrderNotFoundException'
...
   InternalServerError:
    description: Error interno del servidor
    content:
      application/json:
        schema:
          $ref: '#/components/schemas/ErrorResponse'